move compute_control_matrix
