+++ qiskit/extensions/standard/y.py
@@ -20,7 +20,7 @@ from qiskit.circuit import Gate
 from qiskit.circuit import ControlledGate
 from qiskit.circuit import QuantumRegister
 from qiskit.circuit import QuantumCircuit
-from qiskit.extensions._utils import _compute_control_matrix
+from qiskit.circuit._utils import _compute_control_matrix
 from qiskit.qasm import pi
 from qiskit.util import deprecate_arguments
